Fixed build error introduced with r448122 about DataObjectList.append not supporting int type
git-svn-id: https://svn.apache.org/repos/asf/incubator/tuscany/cpp/sca@448123 13f79535-47bb-0310-9956-ffa450edef68

+++ runtime/extensions/ws/reference/axis2c/src/tuscany/sca/ws/Axis2Client.cpp
@@ -292,7 +292,7 @@ namespace tuscany
                                     }
                                 case Operation::INT: 
                                     {
-                                        l.append(*(int*)parm.getValue());
+                                        l.append(*(long*)parm.getValue());
                                         break;
                                     }
                                 case Operation::LONG: 
@@ -307,12 +307,12 @@ namespace tuscany
                                     }
                                 case Operation::UINT: 
                                     {
-                                        l.append(*(unsigned int*)parm.getValue());
+                                        l.append(*(long*)parm.getValue());
                                         break;
                                     }
                                 case Operation::ULONG: 
                                     {
-                                        l.append(*(unsigned long*)parm.getValue());
+                                        l.append(*(long*)parm.getValue());
                                         break;
                                     }
                                 case Operation::FLOAT:
